pyproject.toml
[project]
name = "xfmr-zem"
version = "0.1.0"
description = "Unified Data Pipeline Framework (ZenML + NeMo Curator + DataJuicer) for multi-domain processing"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = "Apache-2.0"
authors = [
    {name = "Khai Hoang", email = "khaihq@vbiacademy.edu.vn"}
]
keywords = ["xfmr-zem", "data-pipeline", "zenml", "nemo-curator", "data-juicer", "mlops"]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "fastmcp>=0.1.0",
    "mcp>=0.1.0",
    "zenml[local,server]>=0.75.0",
]

[project.optional-dependencies]
zenml = [
    "zenml>=0.75.0",
]
nemo = [
    "nemo-curator>=0.6.0",
]
datajuicer = [
    "py-data-juicer>=1.0.0",
]
all = [
    "zenml>=0.75.0",
    "nemo-curator>=0.6.0",
    "py-data-juicer>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
zem = "xfmr_zem.cli:main"
xz = "xfmr_zem.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xfmr_zem"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
